add GameLevel and GameManager
