+++ include/GameManager.cpp
@@ -1,0 +1,5 @@
+#include "GameManager.h"
+
+void GameManager::AddObjectCurrentMap(const Location& loc, AutoGameObject obj)
+{
+}
