+++ test/test.vcxproj
@@ -140,6 +140,7 @@
     <ClInclude Include="..\include\Config_data.h" />
     <ClInclude Include="..\include\Game.h" />
     <ClInclude Include="..\include\GameLevel.h" />
+    <ClInclude Include="..\include\GameManager.h" />
     <ClInclude Include="..\include\GameMap.h" />
     <ClInclude Include="..\include\GameObject.h" />
     <ClInclude Include="..\include\GamePlayer.h" />
@@ -150,6 +151,7 @@
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="..\include\GameLevel.cpp" />
+    <ClCompile Include="..\include\GameManager.cpp" />
     <ClCompile Include="..\src\CollisionManager.cpp" />
     <ClCompile Include="..\src\Config.cpp" />
     <ClCompile Include="..\src\Config_data.cpp" />
